+++ allennlp/data/fields/__init__.py
@@ -5,7 +5,7 @@ that ends up as an array in a model.
 
 from allennlp.data.fields.field import Field
 from allennlp.data.fields.adjacency_field import AdjacencyField
-from allennlp.data.fields.array_field import ArrayField
+from allennlp.data.fields.tensor_field import TensorField
 from allennlp.data.fields.flag_field import FlagField
 from allennlp.data.fields.index_field import IndexField
 from allennlp.data.fields.label_field import LabelField
@@ -17,3 +17,4 @@ from allennlp.data.fields.sequence_field import SequenceField
 from allennlp.data.fields.sequence_label_field import SequenceLabelField
 from allennlp.data.fields.span_field import SpanField
 from allennlp.data.fields.text_field import TextField
+from allennlp.data.fields.array_field import ArrayField
